+++ packages/firestore/test/integration/api/validation.test.ts
@@ -33,11 +33,6 @@ import {
 const FieldPath = firebase.firestore!.FieldPath;
 const FieldValue = firebase.firestore!.FieldValue;
 
-// TODO(b/116617988): Use public API.
-interface FirestoreInternal extends firestore.FirebaseFirestore {
-  _collectionGroup(collectionId: string): firestore.Query;
-}
-
 // We're using 'as any' to pass invalid values to APIs for testing purposes.
 // tslint:disable:no-any
 
@@ -930,8 +925,8 @@ apiDescribe('Validation:', persistence => {
         const query = db
           .collection('collection')
           .orderBy(firebase.firestore!.FieldPath.documentId());
-        const cgQuery = (db as FirestoreInternal)
-          ._collectionGroup('collection')
+        const cgQuery = db
+          .collectionGroup('collection')
           .orderBy(firebase.firestore!.FieldPath.documentId());
         expect(() => query.startAt(1)).to.throw(
           'Invalid query. Expected a string for document ID in ' +
@@ -1054,9 +1049,7 @@ apiDescribe('Validation:', persistence => {
             'FieldPath.documentId(), but it was: 1.'
         );
         expect(() =>
-          (db as FirestoreInternal)
-            ._collectionGroup('foo')
-            .where(FieldPath.documentId(), '>=', 'foo')
+          db.collectionGroup('foo').where(FieldPath.documentId(), '>=', 'foo')
         ).to.throw(
           `Invalid third parameter to Query.where(). When querying a collection group by ` +
             `FieldPath.documentId(), the value provided must result in a valid document path, ` +